refactor: Consolidate authentication and user profile management by removing old controllers and introducing new ones
