--- app/Controllers/Auth/Home.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Controllers\Auth;
-
-use App\Controllers\BaseController;
-
-class Home extends BaseController
-{
-    public function index()
-    {
-        return view('/home');
-    }
-}
